+++ .github/CHANGELOG.md
@@ -52,3 +52,4 @@
 27. [esx_addoninventory]: chore: check for nil society @bitpredator
 28. [npwd]: chore: npwd update to version 1.8.6 @bitpredator
 29. [bpt_deliveries]: chore: clean up the code @bitpredator
+30. [esx-radialmenu]: chore:  fix lint error + clean up the code @bitpredator